Huge Refactor to Konsole Printer. Added colored output
I added a lot in this commit... I still need to go through and clean everything up and do a sanity check for any bugs/refactoring that should take place. I also need to go through everything and identify the todos that I know need to happen throughout. This is growing into much more of a framework that I originallhy thought it would be...